# services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Ocserv User management Api server.",
        "title": "Ocserv User management Example Api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/ocserv/users": {
            "get": {
                "description": "List of Ocserv Users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Users)"
                ],
                "summary": "List of Ocserv Users",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page number, starting from 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Field to order by",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "description": "Sort order, either ASC or DESC",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ocserv_user.OcservUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            },
            "post": {
                "description": "Ocserv User creation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Users)"
                ],
                "summary": "Ocserv User creation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ocserv user create data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ocserv_user.CreateOcservUserData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.OcservUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            }
        },
        "/ocserv/users/ocpasswd": {
            "get": {
                "description": "Ocserv Users from ocpasswd file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Ocpasswd)"
                ],
                "summary": "Ocserv Users from ocpasswd file",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page number, starting from 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Field to order by",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "description": "Sort order, either ASC or DESC",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ocserv_user.OcservUsersSyncResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            }
        },
        "/ocserv/users/ocpasswd/sync": {
            "post": {
                "description": "Ocserv Users from ocpasswd file to db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Ocpasswd)"
                ],
                "summary": "Ocserv Users from ocpasswd file to db",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "list of users with config to sync in db",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ocserv_user.SyncOcpasswdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.OcservUser"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            }
        },
        "/ocserv/users/statistics": {
            "get": {
                "description": "Ocserv Users Statistics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Statistics)"
                ],
                "summary": "Ocserv Users Statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "date_start",
                        "name": "date_start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "date_end",
                        "name": "date_end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DailyTraffic"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            }
        },
        "/ocserv/users/total-bandwidth": {
            "get": {
                "description": "Ocserv Users TotalBandwidth calculating",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Bandwidth)"
                ],
                "summary": "Ocserv Users TotalBandwidth calculating",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "date_start",
                        "name": "date_start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "date_end",
                        "name": "date_end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repository.TotalBandwidths"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "middlewares.Unauthorized": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "ocserv_user.CreateOcservUserData": {
            "type": "object",
            "required": [
                "config",
                "group",
                "password",
                "traffic_type",
                "username"
            ],
            "properties": {
                "config": {
                    "$ref": "#/definitions/models.OcservUserConfig"
                },
                "description": {
                    "type": "string",
                    "maxLength": 1024,
                    "example": "User for testing VPN access"
                },
                "expire_at": {
                    "type": "string",
                    "example": "2025-12-31"
                },
                "group": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                },
                "traffic_size": {
                    "description": "10 GiB",
                    "type": "integer",
                    "minimum": 0,
                    "example": 10737418240
                },
                "traffic_type": {
                    "type": "string",
                    "enum": [
                        "Free",
                        "MonthlyTransmit",
                        "MonthlyReceive",
                        "TotallyTransmit",
                        "TotallyReceive"
                    ],
                    "example": "MonthlyTransmit"
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                }
            }
        },
        "ocserv_user.OcservUsersResponse": {
            "type": "object",
            "required": [
                "meta"
            ],
            "properties": {
                "meta": {
                    "$ref": "#/definitions/request.Meta"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OcservUser"
                    }
                }
            }
        },
...
        "request.ErrorResponse": {
            "type": "object",
            "required": [
                "error",
                "message"
            ],
            "properties": {
                "error": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}